# docs/app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pública para usuários finais (app users) dos tenants. Gerencia registro, autenticação, perfil e catálogo de produtos/serviços.",
        "title": "App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/{url_code}/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Altera a senha do app user autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Alterar senha do app user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Senhas",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/forgot-password": {
            "post": {
                "description": "Envia email com link para reset de senha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Solicitar reset de senha",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.MessageResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/login": {
            "post": {
                "description": "Autentica um app user e retorna token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login de app user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Credenciais",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppAuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalida o token JWT do app user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout de app user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.MessageResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados do app user autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Dados do app user logado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppUserDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/register": {
            "post": {
                "description": "Registra um novo usuário no app do tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar app user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados de registro",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/auth/reset-password": {
            "post": {
                "description": "Reseta a senha usando token recebido por email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Resetar senha",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Token e nova senha",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/catalog/products": {
            "get": {
                "description": "Retorna produtos ativos do tenant paginados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Listar produtos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Itens por página",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/catalog/products/{id}": {
            "get": {
                "description": "Retorna detalhes de um produto ativo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Obter produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ProductResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/catalog/services": {
            "get": {
                "description": "Retorna serviços ativos do tenant paginados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Listar serviços",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Itens por página",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{url_code}/catalog/services/{id}": {
            "get": {
                "description": "Retorna detalhes de um serviço ativo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Obter serviço",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL code do tenant",
                        "name": "url_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ServiceResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_saas-single-db-api_internal_models_swagger.AppAuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                },
                "user": {
                    "$ref": "#/definitions/github_com_saas-single-db-api_internal_models_swagger.AppUserDTO"
                }
            }
        },
        "github_com_saas-single-db-api_internal_models_swagger.AppChangePasswordRequest": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string",
                    "example": "old_password"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "new_password"
                }
            }
        },
        "github_com_saas-single-db-api_internal_models_swagger.AppForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "github_com_saas-single-db-api_internal_models_swagger.AppLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                }
            }
        },
        "github_com_saas-single-db-api_internal_models_swagger.AppRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "secret123"
                }
            }
        },
...
        "github_com_saas-single-db-api_internal_models_swagger.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error_message"
                }
            }
        },
        "github_com_saas-single-db-api_internal_models_swagger.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation_successful"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token de autenticação do app user. Formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}